pdf2world/requirements.txt
attrs==19.3.0
lxml==4.5.0
pdfminer3k==1.3.1
pluggy==0.13.1
ply==3.11
py==1.8.1
pytest==5.4.1
python-docx==0.8.10
six==1.14.0